requirements.txt
# Core Deep Learning
# torch
# torchvision
# segmentation-models-pytorch

# Data Processing & Utils
numpy
pandas
scikit-learn
opencv-python
matplotlib
seaborn
pillow

jupyterlab

streamlit
